add options to preserve files on destination directory

+++ Services/FileManager.php
@@ -109,7 +109,14 @@ class FileManager
                 throw new \Exception("to_folder does not exist");
             }
             $result = null;
-            system("rsync -a --delete " . escapeshellarg($from . '/') . " " . escapeshellarg($to), $result);
+            if (isset($options['preserve_to_files']) && $options['preserve_to_files'])
+            {
+                system("rsync -a" . escapeshellarg($from . '/') . " " . escapeshellarg($to), $result);
+            }
+            else
+            {
+                system("rsync -a --delete " . escapeshellarg($from . '/') . " " . escapeshellarg($to), $result);
+            }
             if ($result !== 0)
             {
                 throw new \Exception("Sync failed");
